requirements.txt
django==4.1.7
python-decouple==3.6
python-dotenv==0.21.0